Updates to doubles from integers

+++ Program.cs
@@ -1,14 +1,14 @@
 ﻿// See https://aka.ms/new-console-template for more information
 //variables
-int num1 = 0;
-int num2 = 0;
-int total = 0;
+double num1 = 0;
+double num2 = 0;
+double total = 0;
 
 //User inputs
 Console.WriteLine("Enter your first number:");
-num1 = Convert.ToInt32(Console.ReadLine());
+num1 = Convert.ToDouble(Console.ReadLine());
 Console.WriteLine("Enter your second number:");
-num2 = Convert.ToInt32(Console.ReadLine());
+num2 = Convert.ToDouble(Console.ReadLine());
 
 //Calculations
 total = num1 + num2;
